+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- Added Telegram control/alerts bot.
+
 ## [0.1.0] - 2024-04-30
 
 ### Added